tests/wpt/tests/css/css-flexbox/animation/order-interpolation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>order interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#order-property">
<meta name="assert" content="order supports animation">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.parent {
  order: 30;
}

.target {
  order: 10;
}
</style>
<body>
<script>
test_interpolation({
  property: 'order',
  from: neutralKeyframe,
  to: '20',
}, [
  {at: -3, expect: '-20'},
  {at: -0.5, expect: '5'},
  {at: 0, expect: '10'},
  {at: 0.3, expect: '13'},
  {at: 0.6, expect: '16'},
  {at: 1, expect: '20'},
  {at: 1.5, expect: '25'},
]);

test_interpolation({
  property: 'order',
  from: 'initial',
  to: '20',
}, [
  {at: -3, expect: '-60'},
  {at: -0.5, expect: '-10'},
  {at: 0, expect: '0'},
  {at: 0.3, expect: '6'},
  {at: 0.6, expect: '12'},
  {at: 1, expect: '20'},
  {at: 1.5, expect: '30'},
]);

test_interpolation({
  property: 'order',
  from: 'inherit',
  to: '20',
}, [
  {at: -3, expect: '60'},
  {at: -0.5, expect: '35'},
  {at: 0, expect: '30'},
  {at: 0.3, expect: '27'},
  {at: 0.6, expect: '24'},
  {at: 1, expect: '20'},
  {at: 1.5, expect: '15'},
]);

test_interpolation({
  property: 'order',
  from: 'unset',
  to: '20',
}, [
  {at: -3, expect: '-60'},
  {at: -0.5, expect: '-10'},
  {at: 0, expect: '0'},
  {at: 0.3, expect: '6'},
  {at: 0.6, expect: '12'},
  {at: 1, expect: '20'},
  {at: 1.5, expect: '30'},
]);

test_interpolation({
  property: 'order',
  from: '10',
  to: '20'
}, [
  {at: -3.0, expect: '-20'},
  {at: -0.5, expect: '5'},
  {at: 0, expect: '10'},
  {at: 0.3, expect: '13'},
  {at: 0.6, expect: '16'},
  {at: 1, expect: '20'},
  {at: 1.5, expect: '25'}
]);

test_interpolation({
  property: 'order',
  from: '2',
  to: '4'
}, [
  {at: -3.0, expect: '-4'},
  {at: -0.5, expect: '1'},
  {at: 0, expect: '2'},
  {at: 0.3, expect: '3'},
  {at: 0.6, expect: '3'},
  {at: 1, expect: '4'},
  {at: 1.5, expect: '5'}
]);
</script>
